Added writing of header
The FileWriter object will now, by default, try to write a header
to the output file if created with mode='w'. We assume that if one
creates a FileWriter object and wants to write to a file then one
typically want a new block. For the future restart option, this is
not what one wants, however this should be handled by a different
mode, like 'a' for append or something similar.