Restructure of plotting
Defined a new object which is used for the plotting. This object
basically just defines a lot of static methods. These static methods
is used for the plotting. Also defined a new method for selecting a
plotting object, this can be expanded in the future to support many
different plotting libraries if this is desirable.