Added PathFile writer.
The new writer is indended for writing out path data in case it's
requested by the user. This is added as a default output for TIS
simulations. It will write out the trajectory and order parameter(s)
for the path. In addition it should write out the energy on the path,
however, this is not added yet. First have to decide how calculated
energies should be stored and which one should be included.