Restructured the output creation.
The output is now given by a name and a type. The type is one of
the pre-defined output types and the name is an unique identifier.
This will make it more difficult to add a completely new type as
one will have to modify createoutput.py. On the other hand it is
now easier to modify the default tasks and add new output tasks
from the predefined ones.

The input settings has been modified to no longer accept an output
keyword, but rather use :output-add: for adding new tasks and
:output-modify: for modifying the existing ones.