Removed fileio sub-package!
This sub-package is now fully replaced by the writers sub-package.
The writers are more homogeneous and file writers can always be
created here. The structure is more general now, which should make
it simpler for the output tasks.
Note that these last changes have probably broken the examples as
they represent large changes to the API. This will be fixed later
when the output task class is fully updated.