Removed several calls to len(path.path)
This is done since I might want to be more flexible in the future.
(The methods should not depend on path.path being something for
which a len() can be called). This is replaced by a simple varialble
path.length which just contains the path length.
Note that this might cause strange behavior if some stupid git
decides to update length outside of path.append().