Restrucutre - moved Path() creation into tis.
integrators.py: Rename the generate_path to a propagate path.
This method will no also require an input path.

tis.py: Made use of the new propagate method. Collected creations
of Path() into this module. The rationale is that the tis methods
should/could choose the path type we want to use.