Updated potentials and simplified them.
The parameters are no longer a property but simple an attribute.
The motivation is that we want a simpler and less comlex code.

For the numpy lennard-jones potential, I have removed the ugly
extra matrices. This means that the code probably runs slower, but
also that it's less hackish and that we use less memory now.
The code is easier to understand after this change.