Changed directories. We will now run in the cwd.
We will now execute the program in the working directory which we
are placed in. This is more predicatable I think. Note that this
will require some gymnastics with directories when we want to import
external libraries. Maybe this can be handled by adding a path to
the environment while executing.