Updated __init__ for SimulationTIS
This is to have the TIS simulation set up in a similar way to
the other simulations, i.e. we are trying to avoid passing the
entire settings-dict down into the object. Here we just pass the
relevant tis-settings and we also require that we pass inn a
PathEnsemble to use for the simulation.