Update docstring for PathEnsembleFile
The docstring should now make it clearer why we have two objects
PathEnsembleWriter and PathEnsembleFile. They are sort of solving
two different problems. PathEnsembleFile is in reality a
PathEnsembleWriter that can be used to load a file and used further
in a path-data analysis in the same way a PathEnsemble object
could be used in the same analysis.