Fixed bug: momentum reset by default in init
The initial path were created with resetting momentum. This is
fixed now. Previously the initial path was generated elsewhere and
not using this method. This bug was discovered by comparing
pathensemble.dat with the code from the previous commit with older
pathensemble.dat files. After this fix, these files should match.