Added methods for merging of .xyz files
This will reverse the backward and merge that with the forward.
Note: No reversal of velocities or check on the number of frames
etc are done. So if the trajectories contain extra frames, this will
be ignored. Also, possible overlaps will be ignored (if the trajectories
come from PyRETIS, they should have no overlaps).