Update index.rst
The MR includes the MR of Raffaela in the latest Master branch.

As from the discussion of Oct 3, 2019, the contribution page shall be discussed once a year and or when a team member change its position.

With the unanimous consent of the work group meeting participants, the main project leader, Titus, will have the final decision on the category in which to fit each person involved (in case of disputes).